Add color & name to player, fix dependencies.
because we're using a shared lib, built by 2 separate gradle builds,
that are each independent, we become responsible for ensuring that
the lib contains only a single version of jackson.

if the 2 gradle builds disagree on jackson level at all, then both
versions end up in the shared lib, and that's a bad thing.

have moved

Signed-off-by: Ozzy Osborne <ozzy@ca.ibm.com>