Propagate location change origin; Offload Kafka
If an origin identifier is recieved on a location change request,
propagate that change in the published event if the change was succcessful.

Create a dedicated thread for publishing events to kafka. Despite Kafka
having its own queue, pushing to that queue was delaying the response
for a successful location change, which was a noticible slowdown in
responsiveness.

Also, the CouchDbConnector should have been @ApplicationScoped, but
wasn't. There should be much less noise in the logs related to connections
to the PlayerDB.

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>